requirements.txt
selenium
webdriver-manager
python-dateutil
Pillow
piexif
pathvalidate
